+++ media_commons_booking_app/package.json
@@ -51,6 +51,7 @@
     "react-bootstrap": "^2.4.0",
     "react-dom": "^18.2.0",
     "react-hook-form": "^7.45.2",
+    "react-router-dom": "^6.22.3",
     "react-transition-group": "^4.4.2",
     "tailwind-datepicker-react": "^1.4.2",
     "tailwindcss": "^3.1.6"
