--- media_commons_booking_app/src/client/admin-page/pa.js
@@ -1,14 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import PAPage from './components/PAPage';
-import './styles.css';
-
-const App = () => {
-  return (
-    <>
-      <PAPage />
-    </>
-  );
-};
-
-ReactDOM.render(<App />, document.getElementById('index'));
