--- media_commons_booking_app/src/client/booking/index.html
@@ -1,17 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <base target="_top" />
-    <!-- Add any external scripts and stylesheets here -->
-    <link
-      href="https://fonts.googleapis.com/css2?family=Nunito"
-      rel="stylesheet"
-    />
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/flowbite/1.8.1/datepicker.min.js"></script>
-  </head>
-  <body>
-    <section id="index">
-      <!-- bundled js and css will get inlined here during build -->
-    </section>
-  </body>
-</html>
