--- media_commons_booking_app/src/client/booking/index.js
@@ -1,14 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import SheetEditor from './components/SheetEditor';
-import './styles.css';
-
-const App = () => {
-  return (
-    <>
-      <SheetEditor />
-    </>
-  );
-};
-
-ReactDOM.render(<App />, document.getElementById('index'));
