+++ media_commons_booking_app/src/client/routes/admin/adminPage.jsx
@@ -1,0 +1,6 @@
+import Admin from './components/Admin';
+import React from 'react';
+
+export default function AdminPage() {
+  return <Admin />;
+}
