+++ media_commons_booking_app/src/client/routes/admin/components/Liaisons.tsx
@@ -1,71 +1,76 @@
-import React, { useState, useEffect } from 'react';
+import {
+  LIAISON_SHEET_NAME,
+  LiaisonType,
+} from '../../booking/components/SheetEditor';
+import React, { useEffect, useState } from 'react';
 
+import { Loading } from '../../../utils/Loading';
+import { formatDate } from '../../../utils/date';
 // This is a wrapper for google.script.run that lets us use promises.
-import { serverFunctions } from '../../utils/serverFunctions';
-import { formatDate } from '../../utils/date';
-import { Loading } from '../../utils/Loading';
+import { serverFunctions } from '../../../utils/serverFunctions';
 
-const PA_USER_SHEET_NAME = 'pa_users';
-
-export type PaUser = {
-  email: string;
-  createdAt: string;
-};
-
-export const PAUsers = () => {
-  const [paUsers, setPaUsers] = useState([]);
-  const [paEmails, setPaEmails] = useState([]);
-  const [mappingPaUsers, setMappingPaUsers] = useState([]);
+export const Liaisons = () => {
+  const [liaisonUsers, setLiaisonUsers] = useState([]);
+  const [liaisonEmails, setAdminEmails] = useState([]);
+  const [mappingLiaisonUsers, setMappingLiaisonUsers] = useState([]);
   const [email, setEmail] = useState('');
+  const [department, setDepartment] = useState('');
 
   useEffect(() => {
-    fetchPaUsers();
+    fetchLiaisonUsers();
   }, []);
   useEffect(() => {
-    const mappings = paUsers
-      .map((paUser, index) => {
+    const mappings = liaisonUsers
+      .map((liaison, index) => {
         if (index !== 0) {
-          return mappingPaUserRows(paUser);
+          return mappingLiaisonRows(liaison);
         }
       })
-      .filter((paUser) => paUser !== undefined);
-    //TODO: filter out paUsers that are not in the future
-    setMappingPaUsers(mappings);
+      .filter((liaison) => liaison !== undefined);
+    //TODO: filter out liaisonUsers that are not in the future
+    setMappingLiaisonUsers(mappings);
     const emails = mappings.map((mapping) => {
       return mapping.email;
     });
-    setPaEmails(emails);
-  }, [paUsers]);
+    setAdminEmails(emails);
+  }, [liaisonUsers]);
 
-  const fetchPaUsers = async () => {
-    serverFunctions.fetchRows(PA_USER_SHEET_NAME).then((rows) => {
-      setPaUsers(rows);
+  const fetchLiaisonUsers = async () => {
+    serverFunctions.fetchRows(LIAISON_SHEET_NAME).then((rows) => {
+      setLiaisonUsers(rows);
     });
   };
 
-  const mappingPaUserRows = (values: string[]): PaUser => {
+  const mappingLiaisonRows = (values: string[]): LiaisonType => {
     return {
       email: values[0],
-      createdAt: values[1],
+      department: values[1],
+      completedAt: values[2],
     };
   };
 
-  const addPaUser = async () => {
+  console.log('liaisonEmails', liaisonEmails);
+  const addLiaisonUser = async () => {
+    if (email === '' || department === '') {
+      alert('Please fill in all the fields');
+      return;
+    }
     setLoading(true);
 
-    if (paEmails.includes(email)) {
+    if (liaisonEmails.includes(email)) {
       alert('This user is already registered');
       return;
     }
 
-    await serverFunctions.appendRow(PA_USER_SHEET_NAME, [
+    await serverFunctions.appendRow(LIAISON_SHEET_NAME, [
       email,
+      department,
       new Date().toString(),
     ]);
 
     alert('User has been registered successfully!');
     setLoading(false);
-    fetchPaUsers();
+    fetchLiaisonUsers();
   };
   const [loading, setLoading] = useState(false);
   if (loading) {
@@ -92,9 +97,30 @@ export const PAUsers = () => {
             required
           />
         </div>
+        <div className="mr-6">
+          <select
+            className="bg-gray-50 border border-gray-300 text-gray-900 text-sm rounded-lg focus:ring-blue-500 focus:border-blue-500 block w-[200px] p-2.5 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-blue-500 dark:focus:border-blue-500"
+            onChange={(e) => {
+              setDepartment(e.target.value);
+            }}
+            value={department}
+          >
+            <option value="" disabled>
+              Select option
+            </option>
+            <option value="ALT">ALT</option>
+            <option value="GameCenter">Game Center</option>
+            <option value="IDM">IDM</option>
+            <option value="ITP / IMA / Low Res">ITP / IMA / Low Res</option>
+            <option value="MARL">MARL</option>
+            <option value="Music Tech">Music Tech</option>
+            <option value="Recorded Music">Recorded Music</option>
+            <option value="others">Other Group</option>
+          </select>
+        </div>
         <button
           type="button"
-          onClick={addPaUser}
+          onClick={addLiaisonUser}
           className="h-[40px] text-white bg-blue-700 hover:bg-blue-800 focus:ring-4 focus:outline-none focus:ring-blue-300 font-medium rounded-lg text-sm w-full sm:w-auto px-5 py-2.5 text-center dark:bg-blue-600 dark:hover:bg-blue-700 dark:focus:ring-blue-800"
         >
           Add User
@@ -109,6 +135,10 @@ export const PAUsers = () => {
                 Email
               </th>
               <th scope="col" className="px-2 py-3">
+                Department
+              </th>
+
+              <th scope="col" className="px-2 py-3">
                 Created Date
               </th>
               <th scope="col" className="px-2 py-3">
@@ -117,16 +147,17 @@ export const PAUsers = () => {
             </tr>
           </thead>
           <tbody>
-            {mappingPaUsers.map((paUser, index) => {
+            {mappingLiaisonUsers.map((liaison, index) => {
               return (
                 <tr
                   key={index}
                   className="bg-white border-b dark:bg-gray-800 dark:border-gray-700 hover:bg-gray-50 dark:hover:bg-gray-600"
                 >
-                  <td className="px-2 py-4 w-36">{paUser.email}</td>
+                  <td className="px-2 py-4 w-36">{liaison.email}</td>
+                  <td className="px-2 py-4 w-36">{liaison.department}</td>
                   <td className="px-2 py-4 w-36">
                     <div className=" flex items-center flex-col">
-                      <div>{formatDate(paUser.createdAt)}</div>{' '}
+                      <div>{formatDate(liaison.completedAt)}</div>{' '}
                     </div>
                   </td>
                   <td className="px-2 py-4 w-36">
@@ -135,12 +166,12 @@ export const PAUsers = () => {
                       onClick={async () => {
                         setLoading(true);
                         await serverFunctions.removeFromList(
-                          PA_USER_SHEET_NAME,
-                          paUser.email
+                          LIAISON_SHEET_NAME,
+                          liaison.email
                         );
                         alert('Successfully removed');
-                        fetchPaUsers();
                         setLoading(false);
+                        fetchLiaisonUsers();
                       }}
                     >
                       Remove
