+++ media_commons_booking_app/src/client/routes/admin/components/SafetyTraining.tsx
@@ -1,71 +1,71 @@
-import React, { useState, useEffect } from 'react';
+import React, { useEffect, useState } from 'react';
 
+import { Loading } from '../../../utils/Loading';
+import { formatDate } from '../../../utils/date';
 // This is a wrapper for google.script.run that lets us use promises.
-import { serverFunctions } from '../../utils/serverFunctions';
-import { formatDate } from '../../utils/date';
-import { Loading } from '../../utils/Loading';
+import { serverFunctions } from '../../../utils/serverFunctions';
 
-export const ADMIN_USER_SHEET_NAME = 'admin_users';
+const SAFETY_TRAINING_SHEET_NAME = 'safety_training_users';
 
-export type AdminUser = {
+type SafetyTraining = {
   email: string;
-  createdAt: string;
+  completedAt: string;
 };
 
-export const AdminUsers = () => {
-  const [adminUsers, setAdminUsers] = useState([]);
-  const [adminEmails, setAdminEmails] = useState([]);
-  const [mappingAdminUsers, setMappingAdminUsers] = useState([]);
+export const SafetyTraining = () => {
+  const [safetyTrainings, setSafetyTrainings] = useState([]);
+  const [trainedEmails, setTrainedEmails] = useState([]);
+  const [mappingTrainings, setMappingTrainings] = useState([]);
   const [email, setEmail] = useState('');
 
   useEffect(() => {
-    fetchAdminUsers();
+    fetchSafetyTrainings();
   }, []);
   useEffect(() => {
-    const mappings = adminUsers
-      .map((adminUser, index) => {
+    const mappings = safetyTrainings
+      .map((safetyTraining, index) => {
         if (index !== 0) {
-          return mappingAdminUserRows(adminUser);
+          return mappingSafetyTrainingRows(safetyTraining);
         }
       })
-      .filter((adminUser) => adminUser !== undefined);
-    //TODO: filter out adminUsers that are not in the future
-    setMappingAdminUsers(mappings);
+      .filter((safetyTraining) => safetyTraining !== undefined);
+    //TODO: filter out safetyTrainings that are not in the future
+    setMappingTrainings(mappings);
     const emails = mappings.map((mapping) => {
       return mapping.email;
     });
-    setAdminEmails(emails);
-  }, [adminUsers]);
+    setTrainedEmails(emails);
+  }, [safetyTrainings]);
 
-  const fetchAdminUsers = async () => {
-    serverFunctions.fetchRows(ADMIN_USER_SHEET_NAME).then((rows) => {
-      setAdminUsers(rows);
+  const fetchSafetyTrainings = async () => {
+    serverFunctions.fetchRows(SAFETY_TRAINING_SHEET_NAME).then((rows) => {
+      setSafetyTrainings(rows);
     });
   };
 
-  const mappingAdminUserRows = (values: string[]): AdminUser => {
+  const mappingSafetyTrainingRows = (values: string[]): SafetyTraining => {
     return {
       email: values[0],
-      createdAt: values[1],
+      completedAt: values[1],
     };
   };
 
-  const addAdminUser = async () => {
+  console.log('trainedEmails', trainedEmails);
+  const addSafetyTrainingUser = async () => {
     setLoading(true);
-
-    if (adminEmails.includes(email)) {
+    if (trainedEmails.includes(email)) {
       alert('This user is already registered');
       return;
     }
 
-    await serverFunctions.appendRow(ADMIN_USER_SHEET_NAME, [
+    await serverFunctions.appendRow(SAFETY_TRAINING_SHEET_NAME, [
       email,
       new Date().toString(),
     ]);
 
     alert('User has been registered successfully!');
     setLoading(false);
-    fetchAdminUsers();
+    fetchSafetyTrainings();
   };
   const [loading, setLoading] = useState(false);
   if (loading) {
@@ -94,7 +94,7 @@ export const AdminUsers = () => {
         </div>
         <button
           type="button"
-          onClick={addAdminUser}
+          onClick={addSafetyTrainingUser}
           className="h-[40px] text-white bg-blue-700 hover:bg-blue-800 focus:ring-4 focus:outline-none focus:ring-blue-300 font-medium rounded-lg text-sm w-full sm:w-auto px-5 py-2.5 text-center dark:bg-blue-600 dark:hover:bg-blue-700 dark:focus:ring-blue-800"
         >
           Add User
@@ -109,24 +109,24 @@ export const AdminUsers = () => {
                 Email
               </th>
               <th scope="col" className="px-2 py-3">
-                Created Date
+                Completed Date
               </th>
               <th scope="col" className="px-2 py-3">
-                Action
+                action
               </th>
             </tr>
           </thead>
           <tbody>
-            {mappingAdminUsers.map((adminUser, index) => {
+            {mappingTrainings.map((safetyTraining, index) => {
               return (
                 <tr
                   key={index}
                   className="bg-white border-b dark:bg-gray-800 dark:border-gray-700 hover:bg-gray-50 dark:hover:bg-gray-600"
                 >
-                  <td className="px-2 py-4 w-36">{adminUser.email}</td>
+                  <td className="px-2 py-4 w-36">{safetyTraining.email}</td>
                   <td className="px-2 py-4 w-36">
                     <div className=" flex items-center flex-col">
-                      <div>{formatDate(adminUser.createdAt)}</div>{' '}
+                      <div>{formatDate(safetyTraining.completedAt)}</div>{' '}
                     </div>
                   </td>
                   <td className="px-2 py-4 w-36">
@@ -135,12 +135,12 @@ export const AdminUsers = () => {
                       onClick={async () => {
                         setLoading(true);
                         await serverFunctions.removeFromList(
-                          ADMIN_USER_SHEET_NAME,
-                          adminUser.email
+                          SAFETY_TRAINING_SHEET_NAME,
+                          safetyTraining.email
                         );
                         alert('Successfully removed');
-                        fetchAdminUsers();
                         setLoading(false);
+                        fetchSafetyTrainings();
                       }}
                     >
                       Remove
