+++ media_commons_booking_app/src/client/routes/booking/bookingPage.jsx
@@ -1,0 +1,6 @@
+import React from 'react';
+import SheetEditor from './components/SheetEditor';
+
+export default function BookingPage() {
+  return <SheetEditor />;
+}
