+++ media_commons_booking_app/src/client/routes/booking/components/Calendars.tsx
@@ -1,9 +1,10 @@
-import React, { useState, useEffect, useRef } from 'react';
+import React, { useEffect, useRef, useState } from 'react';
+import { Role, RoomSetting } from './SheetEditor';
+
 import { CalendarDatePicker } from './CalendarDatePicker';
 import { DateSelectArg } from '@fullcalendar/core';
-import { Role, RoomSetting } from './SheetEditor';
-import { formatDate } from '../../utils/date';
 import { RoomCalendar } from './RoomCalendar';
+import { formatDate } from '../../../utils/date';
 
 type CalendarProps = {
   allRooms: any[];
