+++ media_commons_booking_app/src/client/routes/booking/components/RoleModal.tsx
@@ -21,7 +21,7 @@ export const RoleModal = ({ handleClick }) => {
       data-modal-backdrop="static"
       tabIndex={-1}
       aria-hidden="true"
-      className="fixed top-0 left-0 bg-white right-0 z-50 w-full h-120 p-4 overflow-x-hidden overflow-y-auto md:inset-0 h-[calc(100%-1rem)] max-h-full"
+      className="fixed top-80 left-0 bg-white right-0 z-1 w-full h-120 p-4 overflow-x-hidden overflow-y-auto md:inset-0 h-[calc(100%-1rem)] max-h-full"
     >
       <div className="relative w-full max-w-2xl max-h-full">
         <div className="relative bg-white rounded-lg shadow dark:bg-gray-700">
