+++ media_commons_booking_app/src/client/routes/booking/components/RoomCalendar.tsx
@@ -1,10 +1,11 @@
-import React, { useState, useEffect, useRef } from 'react';
-import interactionPlugin from '@fullcalendar/interaction'; // for selectable
+import React, { useEffect, useRef, useState } from 'react';
+
 import FullCalendar from '@fullcalendar/react';
-import timeGridPlugin from '@fullcalendar/timegrid'; // a plugin!
-import googleCalendarPlugin from '@fullcalendar/google-calendar';
 import dayGridPlugin from '@fullcalendar/daygrid';
-import { serverFunctions } from '../../utils/serverFunctions';
+import googleCalendarPlugin from '@fullcalendar/google-calendar';
+import interactionPlugin from '@fullcalendar/interaction'; // for selectable
+import { serverFunctions } from '../../../utils/serverFunctions';
+import timeGridPlugin from '@fullcalendar/timegrid'; // a plugin!
 
 const TITLE_TAG = '[Click to Delete]';
