+++ media_commons_booking_app/src/client/routes/booking/components/SelectMotionCapture.tsx
