--- media_commons_booking_app/src/client/sidebar-about-page/components/About.jsx
@@ -1,5 +1,0 @@
-import React from 'react';
-
-const About = () => <div>Approved your request!</div>;
-
-export default About;
