--- media_commons_booking_app/src/client/sidebar-about-page/index.html
@@ -1,12 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <base target="_top" />
-    <!-- Add any external scripts and stylesheets here -->
-  </head>
-  <body>
-    <section id="index">
-      <!-- bundled js and css will get inlined here during build -->
-    </section>
-  </body>
-</html>
