--- media_commons_booking_app/src/client/sidebar-about-page/index.js
@@ -1,7 +1,0 @@
-import React from 'react';
-import { createRoot } from 'react-dom/client';
-import About from './components/About';
-
-const container = document.getElementById('index');
-const root = createRoot(container);
-root.render(<About />);
