+++ media_commons_booking_app/src/client/styles.css
@@ -31,3 +31,12 @@ CSSTransitionGroup styling
   opacity: 1;
   transition: opacity 0.5ms ease-in;
 }
+
+.dark {
+  color: white;
+}
+
+.navbar .nav-link.active {
+  background-color: #000;
+  color: #fff;
+}
