+++ media_commons_booking_app/webpack.config.js
@@ -41,32 +41,20 @@ const serverEntry = './src/server/index.ts';
 
 // define appsscript.json file path
 const copyAppscriptEntry = './appsscript.json';
-const copyEmailTemplate = './src/client/booking/approval_email.html';
+const copyEmailTemplate = './src/client/routes/booking/approval_email.html';
 
-const copyApprovalEntry = './src/client/admin-page/approval.html';
-const copyRejectedEntry = './src/client/admin-page/reject.html';
+const copyApprovalEntry = './src/client/routes/approval.html';
+const copyRejectedEntry = './src/client/routes/reject.html';
 // define live development dialog paths
 const devDialogEntry = './dev/index.js';
 
 // define client entry points and output names
 const clientEntrypoints = [
   {
-    name: 'CLIENT - Dialog Demo Tailwind CSS',
-    entry: './src/client/booking/index.js',
-    filename: 'booking',
-    template: './src/client/booking/index.html',
-  },
-  {
-    name: 'CLIENT - Admin',
-    entry: './src/client/admin-page/index.js',
-    filename: 'admin-page',
-    template: './src/client/admin-page/index.html',
-  },
-  {
-    name: 'CLIENT - PA',
-    entry: './src/client/admin-page/pa.js',
-    filename: 'pa-page',
-    template: './src/client/admin-page/index.html',
+    name: 'CLIENT - Media Commons Reservation System',
+    entry: './src/client/index.js',
+    filename: 'index',
+    template: './src/client/index.html',
   },
 ];
